create popupController, replace all its methods to popupController
